# openapi/v1.yaml
openapi: 3.1.0
...
info:
  title: Run Anywhere API
  version: 1.0.0
  summary: Contract-first API for self-hosted Android emulator jobs.
  description: |
    The v1 control-plane contract for projects, signed uploads, Android emulator
    jobs, lifecycle events, artifacts, browser debug sessions, webhooks, workers,
    and runtime profiles.

    Every operation requires an HTTP bearer token. The `x-required-scopes`
    extension records the API-key scope enforced by the future implementation.
  license:
    name: MIT
    identifier: MIT
servers:
  - url: https://api.run-anywhere.example
    description: Example deployment
...
security:
  - BearerAuth: []
...
paths:
...
  /v1/jobs/{job_id}:
    get:
      operationId: getJob
      summary: Get a job
      description: Returns current job state, result, and lifecycle timestamps.
      tags: [Jobs]
      x-required-scopes: [project:read]
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/JobId'
      responses:
        '200':
          description: Job found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
              example:
                id: job_login_flow
                project_id: proj_mobile_qa
                apk_upload_id: upl_app_debug
                runtime_profile: rtp_android_35_pixel_6_x86_64
                mode: headless_ci
                min_isolation: vm_isolated
                automation:
                  type: built_in_smoke
                artifacts:
                  screenshots: true
                  video: false
                  logcat: true
                  junit: true
                timeout_seconds: 900
                state: passed
                outcome: passed
                worker_id: wrk_kvm_01
                created_at: '2026-07-13T08:30:00Z'
                started_at: '2026-07-13T08:30:12Z'
                finished_at: '2026-07-13T08:33:48Z'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/InfrastructureError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API key
      description: Project-scoped API key, or the deployment bootstrap admin token for project creation.
  parameters:
    JobId:
      name: job_id
      in: path
      required: true
      description: Job identifier.
      schema:
        $ref: '#/components/schemas/JobId'
...
  responses:
    ValidationError:
      description: The request is syntactically valid but fails validation.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: validation
              message: timeout_seconds must be positive
              request_id: req_01j2x5h4g8z
              details:
                field: timeout_seconds
    UnauthorizedError:
      description: The bearer token is missing or invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: unauthorized
              message: a valid bearer API key is required
              request_id: req_01j2x5h4g8z
    ForbiddenError:
      description: The credential lacks the required scope or project access.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: forbidden
              message: project:write scope is required
              request_id: req_01j2x5h4g8z
    NotFoundError:
      description: The requested resource does not exist or is not visible to the caller.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: not_found
              message: job was not found
              request_id: req_01j2x5h4g8z
...
    InfrastructureError:
      description: Required infrastructure is temporarily unavailable.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: infra_failed
              message: object storage is unavailable
              request_id: req_01j2x5h4g8z
    InternalError:
      description: An unexpected server error occurred.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: internal_error
              message: an unexpected error occurred
              request_id: req_01j2x5h4g8z
  schemas:
    ProjectId:
      type: string
      pattern: '^proj_[A-Za-z0-9_-]+$'
      example: proj_mobile_qa
    UploadId:
      type: string
      pattern: '^upl_[A-Za-z0-9_-]+$'
      example: upl_app_debug
    RuntimeProfileId:
      type: string
      pattern: '^rtp_[A-Za-z0-9_-]+$'
      example: rtp_android_35_pixel_6_x86_64
    JobId:
      type: string
      pattern: '^job_[A-Za-z0-9_-]+$'
      example: job_login_flow
...
    WorkerId:
      type: string
      pattern: '^wrk_[A-Za-z0-9_-]+$'
      example: wrk_kvm_01
...
    RequestId:
      type: string
      pattern: '^req_[A-Za-z0-9_-]+$'
      example: req_01j2x5h4g8z
...
    ScriptRef:
      type: string
      minLength: 1
      example: s3://mobile-qa/tests/login-flow.zip
    DurationSeconds:
      type: integer
      format: int64
      minimum: 1
      example: 900
...
    IsolationTier:
      type: string
      enum: [vm_isolated, shared_kernel_privileged]
      description: vm_isolated satisfies either minimum; shared_kernel_privileged satisfies only itself.
...
    JobMode:
      type: string
      enum: [headless_ci, browser_debug]
    JobState:
      type: string
      enum:
        - queued
        - claimed
        - provisioning_runtime
        - booting
        - installing_apk
        - running_tests
        - debug_available
        - collecting_artifacts
        - cleaning_up
        - passed
        - failed
        - cancelled
        - timed_out
        - infra_failed
    JobOutcome:
      type: string
      enum: [passed, failed, cancelled, timed_out, infra_failed]
...
    ErrorCode:
      type: string
      enum:
        - validation
        - unauthorized
        - forbidden
        - not_found
        - conflict
        - quota_exceeded
        - infra_failed
        - internal_error
...
    BuiltInSmokeAutomation:
      type: object
      required: [type]
      properties:
        type:
          type: string
          const: built_in_smoke
    AppiumAutomation:
      type: object
      required: [type, script_ref]
      properties:
        type:
          type: string
          const: appium
        script_ref:
          $ref: '#/components/schemas/ScriptRef'
    AutomationSpec:
      oneOf:
        - $ref: '#/components/schemas/BuiltInSmokeAutomation'
        - $ref: '#/components/schemas/AppiumAutomation'
      discriminator:
        propertyName: type
        mapping:
          built_in_smoke: '#/components/schemas/BuiltInSmokeAutomation'
          appium: '#/components/schemas/AppiumAutomation'
      default:
        type: built_in_smoke
    ArtifactSelection:
      type: object
      required: [screenshots, video, logcat, junit]
      properties:
        screenshots:
          type: boolean
        video:
          type: boolean
        logcat:
          type: boolean
        junit:
          type: boolean
...
    FailureDetail:
      type: object
      required: [code, message]
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
          minLength: 1
    Job:
      type: object
      required:
        - id
        - project_id
        - apk_upload_id
        - runtime_profile
        - mode
        - min_isolation
        - automation
        - artifacts
        - timeout_seconds
        - state
        - created_at
      properties:
        id:
          $ref: '#/components/schemas/JobId'
        project_id:
          $ref: '#/components/schemas/ProjectId'
        apk_upload_id:
          $ref: '#/components/schemas/UploadId'
        test_upload_id:
          $ref: '#/components/schemas/UploadId'
        runtime_profile:
          $ref: '#/components/schemas/RuntimeProfileId'
        mode:
          $ref: '#/components/schemas/JobMode'
        min_isolation:
          $ref: '#/components/schemas/IsolationTier'
        automation:
          $ref: '#/components/schemas/AutomationSpec'
        artifacts:
          $ref: '#/components/schemas/ArtifactSelection'
        timeout_seconds:
          $ref: '#/components/schemas/DurationSeconds'
        state:
          $ref: '#/components/schemas/JobState'
        outcome:
          $ref: '#/components/schemas/JobOutcome'
        failure:
          $ref: '#/components/schemas/FailureDetail'
        worker_id:
          $ref: '#/components/schemas/WorkerId'
        created_at:
          type: string
          format: date-time
        started_at:
          type: [string, 'null']
          format: date-time
        finished_at:
          type: [string, 'null']
          format: date-time
...
    ApiError:
      type: object
      required: [code, message, request_id]
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
          minLength: 1
        request_id:
          $ref: '#/components/schemas/RequestId'
        details:
          type: object
          additionalProperties: true
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/ApiError'
...